Web UI Move
- Update website structure where the root contains leaf nodes of each pelican service
