+++ .github/workflows/golangci-lint.yml
@@ -13,7 +13,7 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v3
-      # We embed the contents of src/out/* into the resulting binaries
+      # We embed the contents of web_ui/frontend/out/* into the resulting binaries
       # That particular directory should contain outputs generated by the
       # npm build.  However, to keep the runtime of the linter as fast as possible,
       # instead of running `npm` here, we simply create a dummy empty file.
