+++ origin_ui/origin.go
@@ -19,42 +19,12 @@
 package origin_ui
 
 import (
-	"embed"
-	"mime"
-	"net/http"
-	"os"
-	"path/filepath"
-	"strings"
-
-	"github.com/gin-gonic/gin"
 	"github.com/pelicanplatform/pelican/config"
 	"github.com/pelicanplatform/pelican/param"
 	"github.com/pkg/errors"
+	"os"
+	"path/filepath"
 )
-
-var (
-
-	//go:embed src/out/*
-	webAssets embed.FS
-)
-
-func ConfigOriginUI(engine *gin.Engine) {
-	engine.GET("/view/*path", func(ctx *gin.Context) {
-		path := ctx.Param("path")
-
-		if strings.HasSuffix(path, "/") {
-			path += "index.html"
-		}
-
-		filePath := "src/out" + path
-		file, _ := webAssets.ReadFile(filePath)
-		ctx.Data(
-			http.StatusOK,
-			mime.TypeByExtension(filePath),
-			file,
-		)
-	})
-}
 
 // Configure XrootD directory for both self-based and director-based file transfer tests
 func ConfigureXrootdMonitoringDir() error {
