+++ web_ui/frontend/app/(login)/login/page.tsx
@@ -22,33 +22,36 @@ import {Box, Typography} from "@mui/material";
 import { useRouter } from 'next/navigation'
 import { useState } from "react";
 
-import LoadingButton from "@/app/initialization/code/LoadingButton";
+import LoadingButton from "../components/LoadingButton";
 
-import PasswordInput from "@/app/initialization/password/PasswordInput";
+import PasswordInput from "../components/PasswordInput";
 
 export default function Home() {
 
     const router = useRouter()
-    let [password, _setPassword] = useState <string>("")
-    let [confirmPassword, _setConfirmPassword] = useState <string>("")
+    let [password, setPassword] = useState <string>("")
     let [loading, setLoading] = useState(false);
 
     async function submit(password: string) {
 
         setLoading(true)
 
-        let response = await fetch("/api/v1.0/auth/resetLogin", {
+        let response = await fetch("/api/v1.0/auth/login", {
             method: "POST",
             headers: {
                 "Content-Type": "application/json"
             },
             body: JSON.stringify({
+                "user": "admin",
                 "password": password
             })
         })
 
         if(response.ok){
-            router.push("../../")
+            const url = new URL(window.location.href)
+            const returnURL = url.searchParams.get("returnURL")
+
+            router.push(returnURL ? returnURL : "../")
         } else {
             setLoading(false)
         }
@@ -57,9 +60,7 @@ export default function Home() {
     function onSubmit(e: React.FormEvent<HTMLFormElement>) {
         e.preventDefault()
 
-        if(password == confirmPassword){
-            submit(password)
-        }
+        submit(password)
     }
 
     return (
@@ -67,10 +68,7 @@ export default function Home() {
             <Box m={"auto"} mt={12}  display={"flex"} flexDirection={"column"}>
                 <Box>
                     <Typography textAlign={"center"} variant={"h3"} component={"h3"}>
-                        Set Password
-                    </Typography>
-                    <Typography textAlign={"center"} variant={"h6"} component={"p"}>
-                        Set root metrics password
+                        Login
                     </Typography>
                 </Box>
                 <Box pt={2} mx={"auto"}>
@@ -79,21 +77,9 @@ export default function Home() {
                             <PasswordInput TextFieldProps={{
                                 InputProps: {
                                     onChange: (e) => {
-                                        _setPassword(e.target.value)
+                                        setPassword(e.target.value)
                                     }
                                 }
-                            }}/>
-                        </Box>
-                        <Box>
-                            <PasswordInput TextFieldProps={{
-                                label: "Confirm Password",
-                                InputProps: {
-                                    onChange: (e) => {
-                                        _setConfirmPassword(e.target.value)
-                                    }
-                                },
-                                error: password != confirmPassword,
-                                helperText: password != confirmPassword ? "Passwords do not match" : ""
                             }}/>
                         </Box>
                         <Box mt={3} display={"flex"}>
