+++ web_ui/frontend/app/globals.css
