+++ web_ui/frontend/components/StatusBox.tsx
