+++ web_ui/frontend/components/graphs/LineGraph.tsx
