+++ web_ui/frontend/components/graphs/RateGraph.tsx
