+++ web_ui/frontend/dev/nginx.conf
@@ -29,7 +29,7 @@ http {
             proxy_read_timeout 300s;
             proxy_connect_timeout 10s;
             proxy_set_header X-Real-IP $remote_addr;
-            proxy_pass http://host.docker.internal:3001;
+            proxy_pass http://host.docker.internal:3000;
         }
 
         gzip on;
