+++ web_ui/frontend/next.config.js
@@ -1,10 +1,9 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
     output: process.env.NODE_ENV == "dev" ? "standalone" : 'export',
-    basePath: '/view',
+    basePath: "/view",
     trailingSlash: true,
     images: { unoptimized: true }
 }
-
 
 module.exports = nextConfig
